requirements.txt
Flask
PyMongo
pandas
altair
Pillow
flask_pymongo